requirements.txt
asyncio==3.4.3
attrdict==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
idna==3.4
importlib-metadata==6.0.0
jsonpickle==3.0.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.11
mypy-extensions==0.4.3
numpy==1.21.6
packaging==23.0
Pillow==9.4.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.13
wxasync==0.48
wxPython==4.2.0
zipp==3.11.0
